[wip] Extract text formatting helpers into Modules
Extract a number of ApplicationHelper methods (namely `textilizable`
and its dependencies) into OpenProject::TextFormatting/ObjectLinking
modules. This should enable this functionality to be more easily used
outside of the Rails controller/view context – for instance, from
Grape-based API code.

Also updates WorkPackageModel to include TextFormatting, rather than
ApplicationHelper.

Note: Only RSpecs have been updated thus far.
